// frontend/src/app/models/common.model.ts
import { MatColorName } from "./dash.model";

export interface MftColorScheme {
    tile: MatColorName;
    status: MatColorName;
    flatButton: MatColorName;
    slider: MatColorName;
    valueStatus: MatColorName;
    actionBarIcon: MatColorName;
    typeIcon: MatColorName;
}

export interface MatColorScheme {
    primary : string;
    accent: string;
    warn: string;
    pale: string;
}